Minor refactoring of the tiered storage test harness.
