+++ core/src/test/scala/unit/kafka/utils/RecordsMatcher.scala
@@ -1,11 +1,11 @@
-package integration.kafka.tiered.storage
+package unit.kafka.utils
 
 import java.nio.ByteBuffer
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.record.{Record, SimpleRecord}
-import org.apache.kafka.common.serialization.{Deserializer, Serde, Serdes}
+import org.apache.kafka.common.serialization.{Deserializer, Serde}
 import org.apache.kafka.common.utils.Utils
 import org.hamcrest.{Description, TypeSafeDiagnosingMatcher}
